refactor(personal-assistant): extract gmail, drive, travel, discord domains
Phase B (batch 2): the first cross-domain-dependent domains. Each domain takes
its cross-domain dependencies as a typed deps bundle (cast-free); the delegating
mixin wires them from the composed `this` via a localized cast that is removed
at the final facade-collapse stage. Deps: gmail/drive→google (grant + connector
status), travel→calendar (createCalendarEvent), discord→browser (6 session/tab
methods) + isBrowserPaused.

Plugin typechecks clean in refactor files (only pre-existing app-core unbuilt-
subpath noise in apple-reminders.ts); tests unchanged (105 pass / 4 pre-existing
fails). No behaviour change.
